requirements.txt
# Core dependencies
numpy>=1.19.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Deep learning
torch>=1.13.0
transformers>=4.26.0
sentence-transformers>=2.2.0

# Vector search
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
hnswlib>=0.6.0
annoy>=1.17.0

# Data processing
nltk>=3.6.0
spacy>=3.2.0
datasets>=2.11.0

# Visualization and notebooks
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
hydra-core>=1.1.0
wandb>=0.12.0  # For experiment tracking

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Additional dependencies
model2vec>=0.1.0
joblib>=1.1.0 